Use Command objects to execute commands
Now controllers became actual commands which can be executed with
`execute()` method.  Database, view and configuration are now injected
in command's constructor.  This approach results in a more consistent
way of using commands and cleaner code.  Especially tests became much
cleaner not only because of contructor injections, but also because
dictionaries are used to describe mocks.