Replace existing pages
The user now can scan pages that were already scanned before.
New pages are being inserted in PDF instead of the old ones.
If a notebook has pages, the user should specify which pages he wants to
replace (scan again).  That is achieved by providing a list of page
numbers separated with a space.  Ranges are allowed and can be specify
using dashes.  For example, `1 3-5 4-6` will replace
pages 1, 3, 4, 5, and 6.  An empty input means that no pages should be
replaced.