Add '--create' option for creating new type
When creating new type of a notebook, ask user for type's title and page
size.  Also ask whether pages in notebooks of this type should be
paired.

Fix `create` command. It didn't work because pathlib.Path object was
passed to SQL query where str was expected.  From now path of a notebook
has type of pathlib.Path.  For SQL queries it is converted to str.