Create type and notebook if not exists
When running `scan` command, if no notebooks found in the database,
the `create` command is called automatically, so the user can create a
new notebook immediately.  If no types found in the database when
calling `create` command, the `types --create` command is called
automatically, so the user can create a new type immediately.